[api][webui] Fix repositories migration to fix upgrades to 2.9
This migration would cause upgrades from OBS 2.8 systems relying
on interconnect to break, because many of the projects have
"duplicate" names and labels. This led to a bunch of projects being
filtered out, which then leads to the frontend changing project meta
data for every project that referenced them to change to 'deleted'
and break building packages.